refactor(internal): add repofs for module root and tag dirs
