+++ internal/api.go
@@ -3,11 +3,12 @@ package internal
 import (
 	"bufio"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"log/slog"
 	"os"
 	"path/filepath"
+
+	"github.com/AlexsanderHamir/prof/internal/repofs"
 )
 
 // GetScanner opens a profile file and returns a line scanner and the open file handle.
@@ -22,57 +23,14 @@ func GetScanner(filePath string) (*bufio.Scanner, *os.File, error) {
 	return scanner, file, nil
 }
 
-// CleanOrCreateTag cleans the tag directory if it exists, or creates one.
+// CleanOrCreateTag wraps [repofs.CleanOrCreateTag] using [PermDir].
 func CleanOrCreateTag(dir string) error {
-	info, err := os.Stat(dir)
-	if err != nil {
-		if os.IsNotExist(err) {
-			if err = os.MkdirAll(dir, PermDir); err != nil {
-				return fmt.Errorf("failed to create %s directory: %w", dir, err)
-			}
-			return nil
-		}
-		return err
-	}
-
-	if !info.IsDir() {
-		return fmt.Errorf("path is not a directory: %s", dir)
-	}
-
-	entries, err := os.ReadDir(dir)
-	if err != nil {
-		return fmt.Errorf("failed to read directory: %w", err)
-	}
-
-	for _, entry := range entries {
-		path := filepath.Join(dir, entry.Name())
-		if err = os.RemoveAll(path); err != nil {
-			return fmt.Errorf("failed to remove %s: %w", path, err)
-		}
-	}
-
-	return nil
+	return repofs.CleanOrCreateTag(dir, PermDir)
 }
 
-// FindGoModuleRoot searches upwards from the current working directory for a directory
-// containing a go.mod file and returns its absolute path. If none is found, an error is returned.
+// FindGoModuleRoot forwards to [repofs.FindGoModuleRoot].
 func FindGoModuleRoot() (string, error) {
-	dir, err := os.Getwd()
-	if err != nil {
-		return "", err
-	}
-
-	for {
-		if _, err = os.Stat(filepath.Join(dir, "go.mod")); err == nil {
-			return dir, nil
-		}
-
-		parent := filepath.Dir(dir)
-		if parent == dir {
-			return "", errors.New("go.mod not found from current directory upwards")
-		}
-		dir = parent
-	}
+	return repofs.FindGoModuleRoot()
 }
 
 // PrintConfiguration logs parsed benchmark arguments and optional per-benchmark function filters.
